+++ apps/client/src/page/workspaceSetting/index.tsx
@@ -1,100 +1,46 @@
-import { useEffect, useState } from 'react';
-
 import InfoSetting from '@/page/workspaceSetting/component/InfoSetting';
 import ProfileSetting from '@/page/workspaceSetting/component/ProfileSetting';
 import TeamProfileSetting from '@/page/workspaceSetting/component/TeamProfileSetting';
 import WorkspaceDataSubmitButton from '@/page/workspaceSetting/component/WorkspaceDataSubmitButton/WorkspaceDataSubmitButton';
 import WorkspaceDelete from '@/page/workspaceSetting/component/WorkspaceDelete';
-import { ERROR_NAME, POSITION } from '@/page/workspaceSetting/constant';
-import { usePositionData, useTeamData } from '@/page/workspaceSetting/hook/api/queries';
+import { POSITION } from '@/page/workspaceSetting/constant';
+import { useWorkspaceData } from '@/page/workspaceSetting/hook/useWorkspaceData';
 import { containerStyle } from '@/page/workspaceSetting/styles';
-import { MemberType, TeamType } from '@/page/workspaceSetting/type';
 
 const WorkspaceSettingPage = () => {
-  const [workspaceData, setWorkspaceData] = useState({
-    name: '',
-    position: 'EXECUTIVE',
-    teamName: '',
-    university: '건국대학교',
-    teamIconUrl: '',
-  } as MemberType & Omit<TeamType, 'namingUpdatedAt'>);
-
-  const [initialWorkspaceData, setInitialWorkspaceData] = useState<
-    (MemberType & Omit<TeamType, 'namingUpdatedAt'>) | null
-  >(null);
-
-  const [error, setError] = useState({
-    nameError: ERROR_NAME.VALIDATE,
-    teamNameError: ERROR_NAME.VALIDATE,
-  });
-
-  const { data: positionData } = usePositionData();
-  const { data: teamData } = useTeamData();
-
-  useEffect(() => {
-    if (positionData?.success && teamData?.success && !initialWorkspaceData) {
-      setInitialWorkspaceData({
-        name: positionData.data?.name ?? '',
-        position: positionData.data?.position ?? 'EXECUTIVE',
-        teamName: teamData.data?.teamName ?? '',
-        university: teamData.data?.university ?? '건국대학교',
-        teamIconUrl: teamData.data?.teamIconUrl ?? '',
-      });
-    }
-  }, [positionData, teamData, initialWorkspaceData]);
-
-  useEffect(() => {
-    if (positionData?.success) {
-      setWorkspaceData((prev) => ({
-        ...prev,
-        name: positionData.data?.name ?? '',
-        position: positionData.data?.position ?? 'MEMBER',
-      }));
-    }
-  }, [positionData]);
-
-  useEffect(() => {
-    if (teamData?.success) {
-      setWorkspaceData((prev) => ({
-        ...prev,
-        teamName: teamData?.data?.teamName ?? '',
-        teamIconUrl: teamData?.data?.teamIconUrl ?? '',
-      }));
-    }
-  }, [teamData]);
-
-  const handleWorkspaceDataChange = (key: string, value: string) => {
-    setWorkspaceData((prev) => ({ ...prev, [key]: value }));
-  };
-
-  const handleErrorChange = (key: string, value: string) => {
-    setError((prev) => ({ ...prev, [key]: value }));
-  };
+  const {
+    workspaceData,
+    initialWorkspaceData,
+    handleWorkspaceDataChange,
+    namingUpdatedAt,
+    workspaceDataError,
+    handleWorkspaceDataErrorChange,
+  } = useWorkspaceData();
 
   return (
     <>
       <form css={containerStyle}>
         <WorkspaceDataSubmitButton
           workspaceData={workspaceData}
-          onErrorChange={handleErrorChange}
+          onErrorChange={handleWorkspaceDataErrorChange}
           initialWorkspaceData={initialWorkspaceData}
         />
         <ProfileSetting
           name={workspaceData.name}
           position={workspaceData.position}
           onWorkspaceDataChange={handleWorkspaceDataChange}
-          error={error.nameError}
-          onErrorChange={handleErrorChange}
+          error={workspaceDataError.nameError}
+          onErrorChange={handleWorkspaceDataErrorChange}
         />
 
         {workspaceData.position === POSITION.ADMIN && (
           <>
             <InfoSetting
               teamName={workspaceData.teamName}
-              namingUpdatedAt={teamData?.data?.namingUpdatedAt ?? ''}
+              namingUpdatedAt={namingUpdatedAt ?? ''}
               onWorkspaceDataChange={handleWorkspaceDataChange}
-              error={error.teamNameError}
-              onErrorChange={handleErrorChange}
+              error={workspaceDataError.teamNameError}
+              onErrorChange={handleWorkspaceDataErrorChange}
             />
 
             <TeamProfileSetting
